requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
galaxy.plugin.api==0.67.0
idna==2.10
multidict==5.0.2
typing-extensions==3.7.4.3
yarl==1.6.3
